requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
litellm
chromadb
pymongo
python-dotenv
litellm